utils: replace LazyFormatter with new Formatter
- Replace utils.lazy_formatter with cli.utils.formatter, add the generic
  `format` method and create a Formatter subclass in cli.utils.formatter
  with additional `title` and `filename` methods for usage in cli.main.
- Update HLSStreamWriter.create_decryptor, which was the last remaining
  code using the old LazyFormatter.
