+++ src/streamlink/utils/__init__.py
@@ -5,7 +5,7 @@ from importlib.util import module_from_spec
 from typing import Dict, Generic, Optional, TypeVar
 from urllib.parse import urljoin, urlparse
 
-from streamlink.utils.lazy_formatter import LazyFormatter
+from streamlink.utils.formatter import Formatter
 from streamlink.utils.named_pipe import NamedPipe
 from streamlink.utils.parse import parse_html, parse_json, parse_qsd, parse_xml
 from streamlink.utils.url import update_scheme, url_equal
@@ -134,7 +134,7 @@ __all__ = [
     "absolute_url", "prepend_www",
     "search_dict",
     "LRUCache",
-    "LazyFormatter",
+    "Formatter",
     "NamedPipe",
     "parse_html", "parse_json", "parse_qsd", "parse_xml",
     "update_scheme", "url_equal",
